+++ src/main/java/kitchenpos/domain/TableGroup.java
@@ -2,6 +2,7 @@ package kitchenpos.domain;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import kitchenpos.ordertable.domain.OrderTable;
 
 public class TableGroup {
     private final Long id;
