+++ src/main/java/kitchenpos/ordertable/domain/TableGroupRepository.java
@@ -1,12 +1,12 @@
-package kitchenpos.dao;
+package kitchenpos.ordertable.domain;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.dao.infrastructure.JdbcTemplateOrderTableDao;
+import kitchenpos.dao.TableGroupDao;
 import kitchenpos.dao.infrastructure.JdbcTemplateTableGroupDao;
-import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
+import kitchenpos.ordertable.domain.infra.JdbcTemplateOrderTableDao;
 import org.springframework.stereotype.Repository;
 
 @Repository
