+++ src/test/java/kitchenpos/Fixture.java
@@ -2,7 +2,6 @@ package kitchenpos;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
@@ -10,6 +9,7 @@ import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.product.domain.Product;
 
 @SuppressWarnings("NonAsciiCharacters")
